# swagger.yaml
openapi: 3.0.0
info:
  title: R5 API
  description: Simple REST API
  version: "1.0.0"

servers:
  - url: http://localhost:3000

paths:
  /companies:
    get:
      summary: Get all companies
      responses:
        "200":
          description: A list of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      summary: Add a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Company"
      responses:
        "201":
          description: Company created

  /companies/{id}:
    get:
      summary: Get a company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Company not found

    put:
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Company"
      responses:
        "200":
          description: Company updated
        "404":
          description: Company not found

    delete:
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Company deleted successfully
        "404":
          description: Company not found

components:
  schemas:
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        theme:
          type: string
        description:
          type: string
        details:
          type: string
        meta:
          type: object
          properties:
            description:
              type: string